requirements.txt
opencv-python
numpy
pdf2image
google-genai
pandas
Pillow
ultralytics
python-dotenv
